API configuration moved to another file
