+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "glicocheck-api",
-  "version": "1.0.0.1",
+  "version": "1.0.0.2",
   "description": "API for diabetes management.",
   "main": "index.js",
   "type": "module",
